+++ src/OpenTelemetry.AutoInstrumentation/Instrumentations/MongoDB/Integrations/MongoClientIntegrationExecuteAsync.cs
@@ -4,8 +4,11 @@
 using System.Linq.Expressions;
 using System.Reflection;
 using System.Text.RegularExpressions;
+using OpenTelemetry.AutoInstrumentation;
 using OpenTelemetry.AutoInstrumentation.CallTarget;
-using OpenTelemetry.AutoInstrumentation.Instrumentations.RabbitMq6.DuckTypes;
+using OpenTelemetry.AutoInstrumentation.Instrumentations;
+using OpenTelemetry.AutoInstrumentation.Instrumentations.MongoDB.DuckTypes;
+using OpenTelemetry.AutoInstrumentation.Util;
 
 namespace OpenTelemetry.AutoInstrumentation.Instrumentations.MongoDB.Integrations;
 
@@ -13,10 +16,120 @@ namespace OpenTelemetry.AutoInstrumentation.Instrumentations.MongoDB.Integration
 /// MongoDB.Driver.MongoClient calltarget instrumentation
 /// </summary>
 [InstrumentMethod(
-    assemblyName: MongoDBConstants.AssemblyName,
-    typeName: MongoDBConstants.TypeName,
+    assemblyName: MongoDBConstants.AssemblyName3,
+    typeName: "MongoDB.Driver.Core.WireProtocol.CommandUsingQueryMessageWireProtocol`1",
     methodName: "ExecuteAsync",
-    returnTypeName: "System.Threading.Tasks.Task",
+    returnTypeName: ClrNames.GenericTaskWithGenericClassParameter,
+    parameterTypeNames: new[] { "MongoDB.Driver.Core.Connections.IConnection", ClrNames.CancellationToken },
+    minimumVersion: MongoDBConstants.MinimumVersion3,
+    maximumVersion: MongoDBConstants.MaximumVersion3,
+    integrationName: MongoDBConstants.IntegrationName,
+    type: InstrumentationType.Trace)]
+[InstrumentMethod(
+    assemblyName: MongoDBConstants.AssemblyName3,
+    typeName: "MongoDB.Driver.Core.WireProtocol.CommandUsingCommandMessageWireProtocol`1",
+    methodName: "ExecuteAsync",
+    returnTypeName: ClrNames.GenericTaskWithGenericClassParameter,
+    parameterTypeNames: new[] { "MongoDB.Driver.Core.Connections.IConnection", ClrNames.CancellationToken },
+    minimumVersion: MongoDBConstants.MinimumVersion3,
+    maximumVersion: MongoDBConstants.MaximumVersion3,
+    integrationName: MongoDBConstants.IntegrationName,
+    type: InstrumentationType.Trace)]
+[InstrumentMethod(
+    assemblyName: MongoDBConstants.AssemblyName3,
+    typeName: "MongoDB.Driver.Core.WireProtocol.CommandWireProtocol`1",
+    methodName: "ExecuteAsync",
+    returnTypeName: ClrNames.GenericTaskWithGenericClassParameter,
+    parameterTypeNames: new[] { "MongoDB.Driver.Core.Connections.IConnection", ClrNames.CancellationToken },
+    minimumVersion: MongoDBConstants.MinimumVersion3,
+    maximumVersion: MongoDBConstants.MaximumVersion3,
+    integrationName: MongoDBConstants.IntegrationName,
+    type: InstrumentationType.Trace)]
+[InstrumentMethod(
+    assemblyName: MongoDBConstants.AssemblyName3,
+    typeName: "MongoDB.Driver.Core.WireProtocol.GetMoreWireProtocol`1",
+    methodName: "ExecuteAsync",
+    returnTypeName: ClrNames.GenericTaskWithGenericClassParameter,
+    parameterTypeNames: new[] { "MongoDB.Driver.Core.Connections.IConnection", ClrNames.CancellationToken },
+    minimumVersion: MongoDBConstants.MinimumVersion3,
+    maximumVersion: MongoDBConstants.MaximumVersion3,
+    integrationName: MongoDBConstants.IntegrationName,
+    type: InstrumentationType.Trace)]
+[InstrumentMethod(
+    assemblyName: MongoDBConstants.AssemblyName3,
+    typeName: "MongoDB.Driver.Core.WireProtocol.QueryWireProtocol`1",
+    methodName: "ExecuteAsync",
+    returnTypeName: ClrNames.GenericTaskWithGenericClassParameter,
+    parameterTypeNames: new[] { "MongoDB.Driver.Core.Connections.IConnection", ClrNames.CancellationToken },
+    minimumVersion: MongoDBConstants.MinimumVersion3,
+    maximumVersion: MongoDBConstants.MaximumVersion3,
+    integrationName: MongoDBConstants.IntegrationName,
+    type: InstrumentationType.Trace)]
+[InstrumentMethod(
+    assemblyName: MongoDBConstants.AssemblyName3,
+    typeName: "MongoDB.Driver.Core.WireProtocol.WriteWireProtocolBase`1",
+    methodName: "ExecuteAsync",
+    returnTypeName: ClrNames.GenericTaskWithGenericClassParameter,
+    parameterTypeNames: new[] { "MongoDB.Driver.Core.Connections.IConnection", ClrNames.CancellationToken },
+    minimumVersion: MongoDBConstants.MinimumVersion3,
+    maximumVersion: MongoDBConstants.MaximumVersion3,
+    integrationName: MongoDBConstants.IntegrationName,
+    type: InstrumentationType.Trace)]
+[InstrumentMethod(
+    assemblyName: MongoDBConstants.AssemblyName,
+    typeName: "MongoDB.Driver.Core.WireProtocol.CommandUsingQueryMessageWireProtocol`1",
+    methodName: "ExecuteAsync",
+    returnTypeName: ClrNames.GenericTaskWithGenericClassParameter,
+    parameterTypeNames: new[] { "MongoDB.Driver.Core.Connections.IConnection", ClrNames.CancellationToken },
+    minimumVersion: MongoDBConstants.MinimumVersion,
+    maximumVersion: MongoDBConstants.MaximumVersion,
+    integrationName: MongoDBConstants.IntegrationName,
+    type: InstrumentationType.Trace)]
+[InstrumentMethod(
+    assemblyName: MongoDBConstants.AssemblyName,
+    typeName: "MongoDB.Driver.Core.WireProtocol.CommandUsingCommandMessageWireProtocol`1",
+    methodName: "ExecuteAsync",
+    returnTypeName: ClrNames.GenericTaskWithGenericClassParameter,
+    parameterTypeNames: new[] { "MongoDB.Driver.Core.Connections.IConnection", ClrNames.CancellationToken },
+    minimumVersion: MongoDBConstants.MinimumVersion,
+    maximumVersion: MongoDBConstants.MaximumVersion,
+    integrationName: MongoDBConstants.IntegrationName,
+    type: InstrumentationType.Trace)]
+[InstrumentMethod(
+    assemblyName: MongoDBConstants.AssemblyName,
+    typeName: "MongoDB.Driver.Core.WireProtocol.CommandWireProtocol`1",
+    methodName: "ExecuteAsync",
+    returnTypeName: ClrNames.GenericTaskWithGenericClassParameter,
+    parameterTypeNames: new[] { "MongoDB.Driver.Core.Connections.IConnection", ClrNames.CancellationToken },
+    minimumVersion: MongoDBConstants.MinimumVersion,
+    maximumVersion: MongoDBConstants.MaximumVersion,
+    integrationName: MongoDBConstants.IntegrationName,
+    type: InstrumentationType.Trace)]
+[InstrumentMethod(
+    assemblyName: MongoDBConstants.AssemblyName,
+    typeName: "MongoDB.Driver.Core.WireProtocol.GetMoreWireProtocol`1",
+    methodName: "ExecuteAsync",
+    returnTypeName: ClrNames.GenericTaskWithGenericClassParameter,
+    parameterTypeNames: new[] { "MongoDB.Driver.Core.Connections.IConnection", ClrNames.CancellationToken },
+    minimumVersion: MongoDBConstants.MinimumVersion,
+    maximumVersion: MongoDBConstants.MaximumVersion,
+    integrationName: MongoDBConstants.IntegrationName,
+    type: InstrumentationType.Trace)]
+[InstrumentMethod(
+    assemblyName: MongoDBConstants.AssemblyName,
+    typeName: "MongoDB.Driver.Core.WireProtocol.QueryWireProtocol`1",
+    methodName: "ExecuteAsync",
+    returnTypeName: ClrNames.GenericTaskWithGenericClassParameter,
+    parameterTypeNames: new[] { "MongoDB.Driver.Core.Connections.IConnection", ClrNames.CancellationToken },
+    minimumVersion: MongoDBConstants.MinimumVersion,
+    maximumVersion: MongoDBConstants.MaximumVersion,
+    integrationName: MongoDBConstants.IntegrationName,
+    type: InstrumentationType.Trace)]
+[InstrumentMethod(
+    assemblyName: MongoDBConstants.AssemblyName,
+    typeName: "MongoDB.Driver.Core.WireProtocol.WriteWireProtocolBase`1",
+    methodName: "ExecuteAsync",
+    returnTypeName: ClrNames.GenericTaskWithGenericClassParameter,
     parameterTypeNames: new[] { "MongoDB.Driver.Core.Connections.IConnection", ClrNames.CancellationToken },
     minimumVersion: MongoDBConstants.MinimumVersion,
     maximumVersion: MongoDBConstants.MaximumVersion,
@@ -27,18 +140,25 @@ public static class MongoClientIntegrationExecuteAsync
     internal static CallTargetState OnMethodBegin<TTarget, TConnection>(TTarget instance, TConnection connection, CancellationToken cancellationToken)
         where TConnection : IConnection
     {
-        var scope = MongoDBInstrumentation.StartDatabaseActivity(instance, connection.Endpoint!.HostName, connection.Endpoint!.Port);
-
-        if (scope == null)
-        {
-            return CallTargetState.GetDefault();
-        }
-
-        return new CallTargetState(scope);
+        var activity = MongoDBInstrumentation.StartDatabaseActivity(instance, connection);
+        return new CallTargetState(activity, null);
     }
 
-    internal static CallTargetReturn OnMethodEnd<TTarget>(TTarget instance, Exception exception, in CallTargetState state)
+    internal static TReturn OnAsyncMethodEnd<TTarget, TReturn>(TTarget instance, TReturn returnValue, Exception exception, in CallTargetState state)
     {
-        return CallTargetReturn.GetDefault();
+        var activity = state.Activity;
+
+        if (activity is null)
+        {
+            return returnValue;
+        }
+
+        if (exception is not null)
+        {
+            activity.SetException(exception);
+        }
+
+        activity.Stop();
+        return returnValue;
     }
 }