+++ src/OpenTelemetry.AutoInstrumentation/Instrumentations/MongoDB/MongoDBConstants.cs
@@ -6,9 +6,14 @@ namespace OpenTelemetry.AutoInstrumentation.Instrumentations.MongoDB;
 internal static class MongoDBConstants
 {
     public const string IntegrationName = "MongoDB";
-    public const string AssemblyName = "MongoDB.Driver";
-    public const string TypeName = "MongoDB.Driver.MongoClient";
 
+    // 3.0.0+
+    public const string AssemblyName3 = "MongoDB.Driver";
+    public const string MinimumVersion3 = "3.0.0";
+    public const string MaximumVersion3 = "3.*.*";
+
+    // 2.0.0+
+    public const string AssemblyName = "MongoDB.Driver.Core";
     public const string MinimumVersion = "2.0.0";
-    public const string MaximumVersion = "3.*.*";
+    public const string MaximumVersion = "2.*.*";
 }