+++ src/OpenTelemetry.AutoInstrumentation/Instrumentations/MongoDB/MongoDBInstrumentation.cs
@@ -7,6 +7,7 @@ using System.Diagnostics;
 using System.Globalization;
 using System.Linq;
 using System.Net;
+using System.Net.Sockets;
 using System.Text;
 using System.Threading.Tasks;
 using OpenTelemetry.AutoInstrumentation.DuckTyping;
@@ -20,19 +21,24 @@ internal static class MongoDBInstrumentation
 
     public static Activity? StartDatabaseActivity(
         object? instance,
-        string? serverAddress,
-        int? serverPort,
-        int? batchSize = null,
-        string? statusCode = null)
+        IConnection connection)
     {
         if (instance == null)
         {
             return null;
         }
 
-        var database = GetDatabaseName(instance);
+        if (!TryGetDatabaseName(instance, out var database))
+        {
+            return null;
+        }
 
-        if (!TryGetQueryDetails(instance, database, out var collection, out var operationName))
+        if (!TryGetQueryDetails(instance, out var collection, out var operationName))
+        {
+            return null;
+        }
+
+        if (!TryGetNetworkAttributes(connection, out var serverAddress, out var serverName, out var serverPort))
         {
             return null;
         }
@@ -43,7 +49,7 @@ internal static class MongoDBInstrumentation
 
         if (activity is { IsAllDataRequested: true })
         {
-            SetCommonAttributes(activity, operationName, collection, database, serverAddress, serverPort, batchSize, statusCode);
+            SetCommonAttributes(activity, operationName, collection, database, serverAddress, serverName, serverPort);
         }
 
         return activity;
@@ -55,49 +61,56 @@ internal static class MongoDBInstrumentation
         string? collection,
         string? database,
         string? serverAddress,
-        int? serverPort,
-        int? batchSize,
-        string? statusCode)
+        string? serverName,
+        int? serverPort)
     {
-        activity.SetTag(DatabaseAttributes.Keys.DbSystem, "mongodb");
+        activity.SetTag(DatabaseAttributes.Keys.DbSystem, DatabaseAttributes.Values.MongoDB.MongoDbSystem);
         activity.SetTag(DatabaseAttributes.Keys.DbCollectionName, collection);
         activity.SetTag(DatabaseAttributes.Keys.DbNamespace, database);
         activity.SetTag(DatabaseAttributes.Keys.DbOperationName, operationName);
 
-        activity.SetTag(NetworkAttributes.Keys.ServerAddress, serverAddress);
-        activity.SetTag(NetworkAttributes.Keys.ServerPort, serverPort);
-
-        activity.SetTag(DatabaseAttributes.Keys.DbQueryText, operationName);
-        if (batchSize.HasValue)
+        if (!string.IsNullOrEmpty(serverAddress))
         {
-            activity.SetTag(DatabaseAttributes.Keys.DbOperationBatchSize, batchSize.Value);
+            activity.SetTag(NetworkAttributes.Keys.NetworkPeerAddress, serverAddress);
         }
 
-        if (!string.IsNullOrEmpty(statusCode))
+        if (!string.IsNullOrEmpty(serverName))
         {
-            activity.SetTag(DatabaseAttributes.Keys.DbResponseStatusCode, statusCode);
+            activity.SetTag(NetworkAttributes.Keys.ServerAddress, serverName);
+        }
+
+        if (serverPort is not null)
+        {
+            activity.SetTag(NetworkAttributes.Keys.ServerPort, serverPort);
+            activity.SetTag(NetworkAttributes.Keys.NetworkPeerPort, serverPort);
         }
     }
 
-    private static string? GetDatabaseName(object wireProtocol)
+    private static bool TryGetDatabaseName(object instance, out string? database)
     {
-        if (wireProtocol.TryDuckCast<IWireProtocolWithDatabaseNamespaceStruct>(out var protocolWithDatabaseNamespace)
+        database = null;
+        if (instance.TryDuckCast<IWireProtocolWithDatabaseNamespaceStruct>(out var protocolWithDatabaseNamespace)
          && protocolWithDatabaseNamespace.DatabaseNamespace is not null
          && protocolWithDatabaseNamespace.DatabaseNamespace.TryDuckCast<DatabaseNamespaceStruct>(out var databaseNamespace))
         {
-            return databaseNamespace.DatabaseName;
+            database = databaseNamespace.DatabaseName;
         }
 
-        return null;
+        if (string.IsNullOrEmpty(database))
+        {
+            return false;
+        }
+
+        return true;
     }
 
-    private static bool TryGetQueryDetails(object instance, string? databaseName, out string? collection, out string? operationName)
+    private static bool TryGetQueryDetails(object instance, out string? collection, out string? operationName)
     {
         collection = null;
         operationName = null;
 
         if (instance.TryDuckCast<IWireProtocolWithCommandStruct>(out var protocolWithCommand)
-         && protocolWithCommand.Command != null
+         && protocolWithCommand.Command is not null
          && protocolWithCommand.Command.TryDuckCast<IBsonDocumentProxy>(out var bsonDocument))
         {
             var firstElement = bsonDocument.GetElement(0);
@@ -110,6 +123,44 @@ internal static class MongoDBInstrumentation
 
             collection = firstElement.Value?.ToString();
             operationName = mongoOperationName;
+        }
+
+        if (string.IsNullOrEmpty(collection) || string.IsNullOrEmpty(operationName))
+        {
+            return false;
+        }
+
+        return true;
+    }
+
+    private static bool TryGetNetworkAttributes(IConnection connection, out string? serverAddress, out string? serverName, out int? serverPort)
+    {
+        serverAddress = null;
+        serverName = null;
+        serverPort = null;
+
+        if (connection.EndPoint == null)
+        {
+            return false;
+        }
+
+        if (connection.EndPoint is IPEndPoint ipEndPoint)
+        {
+            serverAddress = ipEndPoint.Address.ToString();
+            serverPort = ipEndPoint.Port;
+
+            serverName = Dns.GetHostEntry(serverAddress).ToString();
+        }
+        else if (connection.EndPoint is DnsEndPoint dnsEndPoint)
+        {
+            serverName = dnsEndPoint.Host;
+            serverPort = dnsEndPoint.Port;
+
+            var ipAddresses = Dns.GetHostAddresses(serverName);
+            if (ipAddresses.Length > 0)
+            {
+                serverAddress = ipAddresses[0].ToString();
+            }
         }
 
         return true;