+++ test/IntegrationTests/LibraryVersions.g.cs
@@ -150,9 +150,11 @@ public static partial class LibraryVersion
 #if DEFAULT_TEST_PACKAGE_VERSIONS
             theoryData.Add(string.Empty);
 #else
-            theoryData.Add("2.28.0");
+            theoryData.Add("2.19.0");
             theoryData.Add("2.30.0");
+#if NET8_0
             theoryData.Add("3.0.0");
+#endif
 #endif
             return theoryData;
         }